+++ cookietemple/lint/domains/lib.py
@@ -9,7 +9,7 @@ class LibCppLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
     def __init__(self, path):
         super().__init__(path)
 
-    def lint(self):
+    def lint(self, skip_external):
         super().lint_project(self, self.methods)
 
     def cpp_files_exist(self) -> None: